requirements.txt
opencv-python-headless==4.9.0.80
Pillow==10.3.0
pytesseract==0.3.10